+++ src/media/capture.rs
@@ -4,10 +4,10 @@ use anyhow::Result;
 use opencv::{prelude::*, videoio};
 use tokio::sync::watch;
 
-pub async fn start_capture(fps: u64, frame_tx: watch::Sender<Mat>) -> Result<()> {
+pub async fn start_capture(fps: f32, frame_tx: watch::Sender<Mat>) -> Result<()> {
     // Open the web-camera (assuming you have one)
     let mut cam = videoio::VideoCapture::new(0, videoio::CAP_ANY)?;
-    let loop_pause_millis = 1000 / fps;
+    let loop_pause_millis = (1000f32 / fps) as u64;
     loop {
         // TODO: Modify frame in place, this method temporarily creates two frames.
         let mut new_frame = Mat::default();
